Employ `add_globals_to_context` in all views (some were missing)
