+++ django/templates/base.html
@@ -100,7 +100,7 @@ function autoscale() {
 </div>
 
 <div id="footer">
-    <a href="https://github.com/kosmotive/cs2pb" target="_blank">
+    <a href="https://github.com/kosmotive/cs2pb/tree/{{ version.branch }}" target="_blank">
         <b>CS2 Performance Board</b>
         {% if version %}
             <br>{{ version.sha|slice:"0:7" }} @ {{ version.date }}
